+++ cachetools/errors.py
@@ -1,0 +1,7 @@
+class ExpiredError(KeyError):
+    """Raised when a cached item's time-to-live has expired.
+
+    This is a subclass of :exc:`KeyError`.
+
+    """
+    pass